+++ rjiter/tests/one_byte_reader.rs
@@ -1,4 +1,4 @@
-use std::io::Read;
+use embedded_io::Read;
 
 pub struct OneByteReader<I>
 where
@@ -16,12 +16,19 @@ where
     }
 }
 
+impl<I> embedded_io::ErrorType for OneByteReader<I>
+where
+    I: Iterator<Item = u8>,
+{
+    type Error = rjiter::error::IoError;
+}
+
 impl<I> Read for OneByteReader<I>
 where
     I: Iterator<Item = u8>,
 {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        if buf.len() == 0 {
+    fn read(&mut self, buf: &mut [u8]) -> Result<usize, Self::Error> {
+        if buf.is_empty() {
             return Ok(0);
         }
         if let Some(next_byte) = self.iter.next() {